06-30 Fixed scripts and sprites
- Removed the "DaTankv2RH0jre" script as "DaTankv2" now works for both
left and right moving tanks

- Removed the "testBox2" script and replaced it with the
"ParashuteTooper" script (which does the same as the other script)

- Added folders for our files (People and Enemies). There is also a
folder for unused files but it should NOT be removed as those files are
used for some of the other files so removing them will FUCK SHIT UP D:

- Added the following sprites to the scene: ParashuteToopers and every
single person. Still missing to add shot effects on tanks and the
bullets (have no sprite for bullets tho).

- Fixed some points in "Steffen Notes" as some of them have now been
solved